+++ flow_collector.php
@@ -44,6 +44,9 @@ ini_set('max_execution_time', '-1');
 
 flowview_connect();
 
+/* include arrays after flowview_connect() */
+include_once($config['base_path'] . '/plugins/flowview/arrays.php');
+
 $debug     = false;
 $reload    = true;
 $taskname  = '';
@@ -1550,48 +1553,52 @@ function process_fv10($p, $ex_addr) {
 }
 
 function flowview_template_supported($template, $tid) {
+	global $required_fields_v4, $required_fields_v6;
+
 	$fieldspec = array('field_id', 'name', 'pack', 'unpack');
 	$columns   = array();
 
-	$required_fields_v4 = array(
-		'octetDeltaCount'             => 1,
-		'packetDeltaCount'            => 2,
-		'protocolIdentifier'          => 4,
-		'ipClassOfService'            => 5,
-		'tcpControlBits'              => 6,
-		'sourceTransportPort'         => 7,
-		'sourceIPv4Address'           => 8,
-		'sourceIPv4PrefixLength'      => 9,
-		'ingressInterface'            => 10,
-		'destinationTransportPort'    => 11,
-		'destinationIPv4Address'      => 12,
-		'destinationIPv4PrefixLength' => 13,
-		'egressInterface'             => 14,
-		'ipNextHopIPv4Address'        => 15,
-		'flowEndSysUpTime'            => 21,
-		'flowStartSysUpTime'          => 22,
-	);
+// These are now global, but leaving here for dev efforts
+//
+//	$required_fields_v4 = array(
+//		'octetDeltaCount'             => 1,
+//		'packetDeltaCount'            => 2,
+//		'protocolIdentifier'          => 4,
+//		'ipClassOfService'            => 5,
+//		'tcpControlBits'              => 6,
+//		'sourceTransportPort'         => 7,
+//		'sourceIPv4Address'           => 8,
+//		'sourceIPv4PrefixLength'      => 9,
+//		'ingressInterface'            => 10,
+//		'destinationTransportPort'    => 11,
+//		'destinationIPv4Address'      => 12,
+//		'destinationIPv4PrefixLength' => 13,
+//		'egressInterface'             => 14,
+//		'ipNextHopIPv4Address'        => 15,
+//		'flowEndSysUpTime'            => 21,
+//		'flowStartSysUpTime'          => 22,
+//	);
 
-	$required_fields_v6 = array(
-		'octetDeltaCount'             => 1,
-		'packetDeltaCount'            => 2,
-		'protocolIdentifier'          => 4,
-		'ipClassOfService'            => 5,
-		'tcpControlBits'              => 6,
-		'sourceTransportPort'         => 7,
-		'ingressInterface'            => 10,
-		'destinationTransportPort'    => 11,
-		'egressInterface'             => 14,
-		'flowEndSysUpTime'            => 21,
-		'flowStartSysUpTime'          => 22,
-		'sourceIPv6Address'           => 27,
-		'destinationIPv6Address'      => 28,
-		'sourceIPv6PrefixLength'      => 29,
-		'destinationIPv6PrefixLength' => 30,
-		'samplingInterval'            => 34,
-		'ipVersion'                   => 60,
-		'ipNextHopIPv6Address'        => 62,
-	);
+//	$required_fields_v6 = array(
+//		'octetDeltaCount'             => 1,
+//		'packetDeltaCount'            => 2,
+//		'protocolIdentifier'          => 4,
+//		'ipClassOfService'            => 5,
+//		'tcpControlBits'              => 6,
+//		'sourceTransportPort'         => 7,
+//		'ingressInterface'            => 10,
+//		'destinationTransportPort'    => 11,
+//		'egressInterface'             => 14,
+//		'flowEndSysUpTime'            => 21,
+//		'flowStartSysUpTime'          => 22,
+//		'sourceIPv6Address'           => 27,
+//		'destinationIPv6Address'      => 28,
+//		'sourceIPv6PrefixLength'      => 29,
+//		'destinationIPv6PrefixLength' => 30,
+//		'samplingInterval'            => 34,
+//		'ipVersion'                   => 60,
+//		'ipNextHopIPv6Address'        => 62,
+//	);
 
 	foreach($template as $index => $field) {
 		$columns[$field['field_id']] = true;
@@ -1619,59 +1626,61 @@ function flowview_template_supported($template, $tid) {
 }
 
 function process_v9_v10($data, $ex_addr, $flowtime, $sysuptime = 0) {
-	global $listener_id, $partition;
+	global $listener_id, $partition, $supported_fields;
 
-	$fieldname = array(
-		'engine_type'       => 38,
-		'engine_id'         => 39,
-		'sampling_interval' => 34,
-		'ipVersion'         => 60,
-		'sysuptime'         => 160,
-
-		'src_addr'          => 8,
-		'src_addr_ipv6'     => 27,
-		'src_prefix'        => 9,
-		'src_prefix_ipv6'   => 29,
-		'src_if'            => 10,
-		'src_as'            => 16,
-
-		'dst_addr'          => 12,
-		'dst_addr_ipv6'     => 28,
-		'dst_prefix'        => 13,
-		'dst_prefix_ipv6'   => 30,
-		'dst_if'            => 14,
-		'dst_as'            => 17,
-
-		'nexthop'           => 15,
-		'nexthop_ipv6'      => 62,
-
-		'dPkts'             => 2,
-		'dOctets'           => 1,
-		'src_port'          => 7,
-		'dst_port'          => 11,
-
-		'protocol'          => 4,
-		'tos'               => 5,
-		'flags'             => 6,
-		'start_time'        => 22,
-		'end_time'          => 21
-	);
+//  These are now global, but keeping these here for development
+//
+//	$supported_fields = array(
+//		'engine_type'       => 38,
+//		'engine_id'         => 39,
+//		'sampling_interval' => 34,
+//		'ipVersion'         => 60,
+//		'sysuptime'         => 160,
+//
+//		'src_addr'          => 8,
+//		'src_addr_ipv6'     => 27,
+//		'src_prefix'        => 9,
+//		'src_prefix_ipv6'   => 29,
+//		'src_if'            => 10,
+//		'src_as'            => 16,
+//
+//		'dst_addr'          => 12,
+//		'dst_addr_ipv6'     => 28,
+//		'dst_prefix'        => 13,
+//		'dst_prefix_ipv6'   => 30,
+//		'dst_if'            => 14,
+//		'dst_as'            => 17,
+//
+//		'nexthop'           => 15,
+//		'nexthop_ipv6'      => 62,
+//
+//		'dPkts'             => 2,
+//		'dOctets'           => 1,
+//		'src_port'          => 7,
+//		'dst_port'          => 11,
+//
+//		'protocol'          => 4,
+//		'tos'               => 5,
+//		'flags'             => 6,
+//		'start_time'        => 22,
+//		'end_time'          => 21
+//	);
 
 	$flows = 1;
 
-	if (isset($data[$fieldname['src_addr_ipv6']])) {
-		$src_addr   = $data[$fieldname['src_addr_ipv6']];
+	if (isset($data[$supported_fields['src_addr_ipv6']])) {
+		$src_addr   = $data[$supported_fields['src_addr_ipv6']];
 
-		if (isset($data[$fieldname['src_prefix_ipv6']])) {
-			$src_prefix = $data[$fieldname['src_prefix_ipv6']];
+		if (isset($data[$supported_fields['src_prefix_ipv6']])) {
+			$src_prefix = $data[$supported_fields['src_prefix_ipv6']];
 		} else {
 			$src_prefix = 0;
 		}
-	} elseif (isset($data[$fieldname['src_addr']])) {
-		$src_addr   = $data[$fieldname['src_addr']];
+	} elseif (isset($data[$supported_fields['src_addr']])) {
+		$src_addr   = $data[$supported_fields['src_addr']];
 
-		if (isset($data[$fieldname['src_prefix']])) {
-			$src_prefix = $data[$fieldname['src_prefix']];
+		if (isset($data[$supported_fields['src_prefix']])) {
+			$src_prefix = $data[$supported_fields['src_prefix']];
 		} else {
 			$src_prefix = 0;
 		}
@@ -1680,19 +1689,19 @@ function process_v9_v10($data, $ex_addr, $flowtime, $sysuptime = 0) {
 		return false;
 	}
 
-	if (isset($data[$fieldname['dst_addr_ipv6']])) {
-		$dst_addr = $data[$fieldname['dst_addr_ipv6']];
+	if (isset($data[$supported_fields['dst_addr_ipv6']])) {
+		$dst_addr = $data[$supported_fields['dst_addr_ipv6']];
 
-		if (isset($data[$fieldname['dst_prefix_ipv6']])) {
-			$dst_prefix = $data[$fieldname['dst_prefix_ipv6']];
+		if (isset($data[$supported_fields['dst_prefix_ipv6']])) {
+			$dst_prefix = $data[$supported_fields['dst_prefix_ipv6']];
 		} else {
 			$dst_prefix = 0;
 		}
-	} elseif (isset($data[$fieldname['dst_addr']])) {
-		$dst_addr = $data[$fieldname['dst_addr']];
+	} elseif (isset($data[$supported_fields['dst_addr']])) {
+		$dst_addr = $data[$supported_fields['dst_addr']];
 
-		if (isset($data[$fieldname['dst_prefix']])) {
-			$dst_prefix = $data[$fieldname['dst_prefix']];
+		if (isset($data[$supported_fields['dst_prefix']])) {
+			$dst_prefix = $data[$supported_fields['dst_prefix']];
 		} else {
 			$dst_prefix = 0;
 		}
@@ -1701,46 +1710,46 @@ function process_v9_v10($data, $ex_addr, $flowtime, $sysuptime = 0) {
 		return false;
 	}
 
-	if (isset($data[$fieldname['nexthop_ipv6']])) {
-		$nexthop = $data[$fieldname['nexthop_ipv6']];
-	} elseif (isset($data[$fieldname['nexthop']])) {
-		$nexthop = $data[$fieldname['nexthop']];
+	if (isset($data[$supported_fields['nexthop_ipv6']])) {
+		$nexthop = $data[$supported_fields['nexthop_ipv6']];
+	} elseif (isset($data[$supported_fields['nexthop']])) {
+		$nexthop = $data[$supported_fields['nexthop']];
 	} else {
 		$nexthop = '';
 	}
 
-	if (isset($data[$fieldname['sysuptime']]) && abs($data[$fieldname['end_time']] - $data[$fieldname['sysuptime']]) < 3) {
-		$rstime = ($data[$fieldname['start_time']] - $data[$fieldname['sysuptime']]) / 1000;
-		$rsmsec = substr($data[$fieldname['start_time']] - $data[$fieldname['sysuptime']], -3);
-		$retime = ($data[$fieldname['end_time']] - $data[$fieldname['sysuptime']]) / 1000;
-		$remsec = substr($data[$fieldname['end_time']] - $data[$fieldname['sysuptime']], -3);
+	if (isset($data[$supported_fields['sysuptime']]) && abs($data[$supported_fields['end_time']] - $data[$supported_fields['sysuptime']]) < 3) {
+		$rstime = ($data[$supported_fields['start_time']] - $data[$supported_fields['sysuptime']]) / 1000;
+		$rsmsec = substr($data[$supported_fields['start_time']] - $data[$supported_fields['sysuptime']], -3);
+		$retime = ($data[$supported_fields['end_time']] - $data[$supported_fields['sysuptime']]) / 1000;
+		$remsec = substr($data[$supported_fields['end_time']] - $data[$supported_fields['sysuptime']], -3);
 
 		$start_time = date('Y-m-d H:i:s.v', intval($flowtime + $rstime)) . '.' . $rsmsec;
 		$end_time   = date('Y-m-d H:i:s.v', intval($flowtime + $retime)) . '.' . $remsec;
-		$sysuptime = $data[$fieldname['sysuptime']];
+		$sysuptime = $data[$supported_fields['sysuptime']];
 
-		//debug("Flow: Case 3 SysUptime:{$sysuptime}, StartTime:{$data[$fieldname['start_time']]}, StartDate:{$start_date}, EndTime:{$data[$fieldname['end_time']]}, EndDate:{$end_date}");
+		//debug("Flow: Case 3 SysUptime:{$sysuptime}, StartTime:{$data[$supported_fields['start_time']]}, StartDate:{$start_date}, EndTime:{$data[$supported_fields['end_time']]}, EndDate:{$end_date}");
 	} elseif ($sysuptime > 0) {
 		$rsmsec = $rstime = $remsec = $retime = 0;
 
-		if (isset($data[$fieldname['start_time']])) {
-			$rstime = ($data[$fieldname['start_time']] - $sysuptime) / 1000;
-			$rsmsec = substr('000' . ($data[$fieldname['start_time']] - $sysuptime), -3);
+		if (isset($data[$supported_fields['start_time']])) {
+			$rstime = ($data[$supported_fields['start_time']] - $sysuptime) / 1000;
+			$rsmsec = substr('000' . ($data[$supported_fields['start_time']] - $sysuptime), -3);
 		}
 
-		if (isset($data[$fieldname['end_time']])) {
-			$retime = ($data[$fieldname['end_time']] - $sysuptime) / 1000;
-			$remsec = substr('000' . ($data[$fieldname['end_time']] - $sysuptime), -3);
+		if (isset($data[$supported_fields['end_time']])) {
+			$retime = ($data[$supported_fields['end_time']] - $sysuptime) / 1000;
+			$remsec = substr('000' . ($data[$supported_fields['end_time']] - $sysuptime), -3);
 		}
 
 		$start_date = date('Y-m-d H:i:s', intval($flowtime + $rstime)) . '.' . $rsmsec;
 		$end_date   = date('Y-m-d H:i:s', intval($flowtime + $retime)) . '.' . $remsec;
 
-		//debug("Flow: Case 2 SysUptime:{$sysuptime}, StartTime:{$data[$fieldname['start_time']]}, StartDate:{$start_date}, EndTime:{$data[$fieldname['end_time']]}, EndDate:{$end_date}");
+		//debug("Flow: Case 2 SysUptime:{$sysuptime}, StartTime:{$data[$supported_fields['start_time']]}, StartDate:{$start_date}, EndTime:{$data[$supported_fields['end_time']]}, EndDate:{$end_date}");
 	} else {
-		if (isset($data[$fieldname['start_time']]) && isset($data[$fieldname['end_time']])) {
-			$delta_milli = intval(($data[$fieldname['end_time']] - $data[$fieldname['start_time']]) / 1000);
-			$delta_sec   = floor($data[$fieldname['end_time']] - $data[$fieldname['start_time']]);
+		if (isset($data[$supported_fields['start_time']]) && isset($data[$supported_fields['end_time']])) {
+			$delta_milli = intval(($data[$supported_fields['end_time']] - $data[$supported_fields['start_time']]) / 1000);
+			$delta_sec   = floor($data[$supported_fields['end_time']] - $data[$supported_fields['start_time']]);
 		} else {
 			$delta_milli = $delta_sec = 0;
 		}
@@ -1757,61 +1766,61 @@ function process_v9_v10($data, $ex_addr, $flowtime, $sysuptime = 0) {
 	$dst_domain  = flowview_get_dns_from_ip($dst_addr, 100);
 	$dst_rdomain = flowview_get_rdomain_from_domain($dst_domain, $dst_addr);
 
-	if (isset($data[$fieldname['src_port']])) {
-		$src_rport = flowview_translate_port($data[$fieldname['src_port']], false, false);
+	if (isset($data[$supported_fields['src_port']])) {
+		$src_rport = flowview_translate_port($data[$supported_fields['src_port']], false, false);
 	} else {
 		$src_rport = 0;
 	}
 
-	if (isset($data[$fieldname['dst_port']])) {
-		$dst_rport = flowview_translate_port($data[$fieldname['dst_port']], false, false);
+	if (isset($data[$supported_fields['dst_port']])) {
+		$dst_rport = flowview_translate_port($data[$supported_fields['dst_port']], false, false);
 	} else {
 		$dst_rport = 0;
 	}
 
-	if (isset($data[$fieldname['dPkts']]) && $data[$fieldname['dPkts']] > 0) {
-		$pps = round($data[$fieldname['dOctets']] / $data[$fieldname['dPkts']], 3);
+	if (isset($data[$supported_fields['dPkts']]) && $data[$supported_fields['dPkts']] > 0) {
+		$pps = round($data[$supported_fields['dOctets']] / $data[$supported_fields['dPkts']], 3);
 	} else {
 		$pps = 0;
 	}
 
 	$sql = '(' .
 		$listener_id                                      . ', ' .
-		check_set($data, $fieldname['engine_type'])       . ', ' .
-		check_set($data, $fieldname['engine_id'])         . ', ' .
-		check_set($data, $fieldname['sampling_interval']) . ', ' .
+		check_set($data, $supported_fields['engine_type'])       . ', ' .
+		check_set($data, $supported_fields['engine_id'])         . ', ' .
+		check_set($data, $supported_fields['sampling_interval']) . ', ' .
 		db_qstr($ex_addr)                                 . ', ' .
 		$sysuptime                                        . ', ' .
 
 		'INET6_ATON("' . $src_addr . '")'                 . ', ' .
 		db_qstr($src_domain)                              . ', ' .
 		db_qstr($src_rdomain)                             . ', ' .
-		check_set($data, $fieldname['src_as'])            . ', ' .
-		check_set($data, $fieldname['src_if'])            . ', ' .
+		check_set($data, $supported_fields['src_as'])            . ', ' .
+		check_set($data, $supported_fields['src_if'])            . ', ' .
 		$src_prefix                                       . ', ' .
-		check_set($data, $fieldname['src_port'])          . ', ' .
+		check_set($data, $supported_fields['src_port'])          . ', ' .
 		db_qstr($src_rport)                               . ', ' .
 
 		'INET6_ATON("' . $dst_addr . '")'                 . ', ' .
 		db_qstr($dst_domain)                              . ', ' .
 		db_qstr($dst_rdomain)                             . ', ' .
-		check_set($data, $fieldname['dst_as'])            . ', ' .
-		check_set($data, $fieldname['dst_if'])            . ', ' .
+		check_set($data, $supported_fields['dst_as'])            . ', ' .
+		check_set($data, $supported_fields['dst_if'])            . ', ' .
 		$dst_prefix                                       . ', ' .
-		check_set($data, $fieldname['dst_port'])          . ', ' .
+		check_set($data, $supported_fields['dst_port'])          . ', ' .
 		db_qstr($dst_rport)                               . ', ' .
 
 		db_qstr($nexthop)                                 . ', ' .
-		check_set($data, $fieldname['protocol'])          . ', ' .
+		check_set($data, $supported_fields['protocol'])          . ', ' .
 		db_qstr($start_date)                              . ', ' .
 		db_qstr($end_date)                                . ', ' .
 
 		$flows                                            . ', ' .
-		check_set($data, $fieldname['dPkts'])             . ', ' .
-		check_set($data, $fieldname['dOctets'])           . ', ' .
+		check_set($data, $supported_fields['dPkts'])             . ', ' .
+		check_set($data, $supported_fields['dOctets'])           . ', ' .
 		$pps                                              . ', ' .
-		check_set($data, $fieldname['tos'])               . ', ' .
-		check_set($data, $fieldname['flags'])             . ')';
+		check_set($data, $supported_fields['tos'])               . ', ' .
+		check_set($data, $supported_fields['flags'])             . ')';
 
 	return $sql;
 }