+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 * feature: Add support for Net/DNS2 to improve alternate DNS resolution support
 * feature: Allow DNS Cache to be purged with a single button press.
 * feature: Append local domain to locally resolved addresses from the system DNS
+* feature: Show supported or not supported status when viewing Flow Templates
 * feature#9: Support Pie Charts Instead of Bar
 * feature#97: Add FreeBSD service control
 * feature#112: Provide a means to view the DNS Cache and Prune/Edit Entries